pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "allianceauth-blacklist"
dynamic = ["version", "description" ]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
authors = [
    { name = "AaronKable", email = "aaronkable@gmail.com" },
]
keywords = [
    "allianceauth",
    "eveonline",
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Celery",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content"
]
dependencies = [
    "allianceauth>=3.0.0,<5.0.0",
    "djangorestframework",
    "django-filter",
]

[project.optional-dependencies]
test = [
    "allianceauth-securegroups"
]

[project.urls]
Homepage = "https://github.com/Solar-Helix-Independent-Transport/allianceauth-blacklist"
Source = "https://github.com/Solar-Helix-Independent-Transport/allianceauth-blacklist"
Tracker = "https://github.com/Solar-Helix-Independent-Transport/allianceauth-blacklist/issues"

[tool.flit.module]
name = "blacklist"

[tool.isort]
profile = "django"
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "ALLIANCEAUTH",
    "FIRSTPARTY",
    "LOCALFOLDER"
]
known_allianceauth = ["allianceauth", "esi"]
known_django = ["django"]
skip_gitignore = true

[tool.flake8]
exclude = [".git", "*migrations*", ".tox", "dist", "htmlcov"]
max-line-length = 119
select = ["C", "E", "F", "W", "B", "B950"]
ignore = ['E203', 'E231', 'E501', 'W503', 'W291', 'W293']

[tool.djlint]
max_attribute_length=119
max_line_length=119
max_blank_lines=1
